+++ src/react/src/util/constants.jsx
@@ -29,8 +29,9 @@ export const EMPTY_PLACEHOLDER = 'N/A';
 export const InfoPaths = {
     storiesResources: 'stories-resources',
     privacyPolicy: 'privacy-policy',
-    dataCleaningService: 'data-cleaning-service',
+    spotlight: 'spotlight',
     preparingData: 'resources/preparing-data',
+    dataCleaningService: 'data-cleaning-service',
     dataQuality: 'resources/a-free-universal-id-matching-algorithm',
     claimedFacilities: 'stories-resources/claim-a-facility',
     termsOfService: 'terms-of-service',
@@ -868,7 +869,7 @@ export const NavbarItems = [
                     ],
                 },
                 {
-                    label: 'Premium Features',
+                    label: 'Featured Solutions',
                     items: [
                         {
                             type: 'link',
@@ -877,13 +878,13 @@ export const NavbarItems = [
                         },
                         {
                             type: 'link',
-                            label: 'Data Cleaning Service',
-                            href: `${InfoLink}/${InfoPaths.dataCleaningService}`,
+                            label: 'Embedded Map',
+                            href: `${InfoLink}/${InfoPaths.embeddedMap}`,
                         },
                         {
                             type: 'link',
-                            label: 'Embedded Map',
-                            href: `${InfoLink}/${InfoPaths.embeddedMap}`,
+                            label: 'Spotlight',
+                            href: `${InfoLink}/${InfoPaths.spotlight}`,
                         },
                     ],
                 },
@@ -969,7 +970,7 @@ export const NavbarItems = [
     },
     {
         type: 'link',
-        label: 'Pricing',
+        label: 'Solutions',
         href: `${InfoLink}/${InfoPaths.pricing}`,
     },
     { type: 'international' },